TLS support
This introduces basic TLS support.  The server can optionally be
configured to allow TLS by returning an appropriate TlsConfig object
from the shim.

The TLS is handled during the handshake, and is achieved by replacing the
`Read` and `Write` instances on PacketReader and PacketWriter with TLS
wrapped Read and Write implementations.