Fix handshake bug when SSL is enabled
This fixes a bug where the username was not being populated in the
`ClientHandshake` struct when SSL is enabled.

When the client requests SSL (and so CapabilityFlags::CLIENT_SSL is set)
the protocol expects to stop the handshake before the username is sent
by the client, switch to SSL and then restart the handshake (this time
including SSL) but over SSL instead of plain text.

This bug was introduced because the username was _always_ not being read
from the data sent by the server when SSL is enabled, instead of
skipping it the first time, but reading it as expected when restarting
the handshake over SSL.

This change fixes the bug, by passing in extra information to the client
handshake function.  Additional tests are also added to cover this
logic.