Switch to current thread runtime
This switches to use the current thread runtime in tests, to avoid
spawning additional threads.

An interesting side effect of this change alone is that the tests hang
waiting for the joinhandle of the `MysqlIntermediary::run_on_tcp`
thread, which never returns.  This seems to be because the runtime
hasn't finished cleaning up by the time `block_on` exits, so the tcp
connection remains open. Calling `rt.shutdown_background` immediately
after `rt.block_on` fixes this.