Fix buffering during switch to TLS
This fixes a race bug where the intereaction between our buffering and
the TLS handshake can cause problems.

There was an assert in the code to ensure we don't read "too far" into
the stream before we enter the TLS handshake.  However, it turns out
that sometimes (when the client sends the tls handshake information
quickly enough after the pre-tls data) we can end up hitting this
assertion.

This adds a test for this scenario, by using a new `DelayedReadRW` wrapper
in the server which reliably triggers this.

The fix introduces a new `PrependedReader` type that allows us to push
any extra data we may have buffered into it (together with the
underlying connection) before passing it in to rustls for the handshake.